python-backend/requirements.txt
fastapi
uvicorn[standard]
ultralytics
opencv-python-headless
numpy
firebase-admin
httpx
python-multipart
python-jose[cryptography]
python-dotenv
